Modify AttractionModel.py
